# contrato/openapi.json
{
  "components": {
    "schemas": {
...
      "ChatDto": {
        "properties": {
          "confirmacao": {
            "$ref": "#/components/schemas/ConfirmacaoDto"
          },
          "conversaId": {
            "maxLength": 100,
            "type": "string"
          },
          "mensagem": {
            "maxLength": 8000,
            "type": "string"
          },
          "modo": {
            "enum": [
              "assistido",
              "autopiloto"
            ],
            "type": "string"
          },
          "oportunidadeId": {
            "maxLength": 100,
            "type": "string"
          }
        },
        "type": "object"
      },
      "ConfirmacaoDto": {
        "properties": {
          "ajustes": {
            "additionalProperties": true,
            "type": "object"
          },
          "callId": {
            "maxLength": 100,
            "type": "string"
          },
          "decisao": {
            "enum": [
              "confirmar",
              "recusar"
            ],
            "type": "string"
          }
        },
        "required": [
          "callId",
          "decisao"
        ],
        "type": "object"
      },
...
      "MensagemRecrutadorDto": {
        "properties": {
          "contexto": {
            "maxLength": 8000,
            "type": "string"
          },
          "oportunidadeId": {
            "maxLength": 100,
            "type": "string"
          }
        },
        "required": [
          "oportunidadeId"
        ],
        "type": "object"
      },
...
      "RespostaObjeto18Dto": {
        "properties": {
          "loteId": {
            "type": "string"
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "loteId",
          "total"
        ],
        "type": "object"
      },
      "RespostaObjeto19Dto": {
        "properties": {
          "loteId": {
            "type": "string"
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "loteId",
          "total"
        ],
        "type": "object"
      },
...
      "RespostaObjeto20Dto": {
        "properties": {
          "disponivel": {
            "enum": [
              false,
              true
            ],
            "type": "boolean"
          },
          "documentos": {
            "type": "number"
          },
          "porOrigem": {
            "$ref": "#/components/schemas/RespostaObjeto21Dto"
          },
          "ultimaIndexacao": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "documentos",
          "ultimaIndexacao",
          "porOrigem",
          "disponivel"
        ],
        "type": "object"
      },
      "RespostaObjeto21Dto": {
        "properties": {
          "candidatura": {
            "type": "number"
          },
          "nota": {
            "type": "number"
          },
          "perfil": {
            "type": "number"
          }
        },
        "required": [
          "perfil",
          "candidatura",
          "nota"
        ],
        "type": "object"
      },
      "RespostaObjeto22Dto": {
        "properties": {
          "atualizadoEm": {
            "format": "date-time",
            "type": "string"
          },
          "criadoEm": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "modo": {
            "enum": [
              "assistido",
              "autopiloto"
            ],
            "type": "string"
          },
          "oportunidadeId": {
            "nullable": true,
            "type": "string"
          },
          "titulo": {
            "type": "string"
          },
          "totalMensagens": {
            "type": "number"
          },
          "ultimaMensagem": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "modo",
          "oportunidadeId",
          "titulo",
          "ultimaMensagem",
          "totalMensagens",
          "criadoEm",
          "atualizadoEm"
        ],
        "type": "object"
      },
      "RespostaObjeto23Dto": {
        "properties": {
          "atualizadoEm": {
            "format": "date-time",
            "type": "string"
          },
          "criadoEm": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "mensagens": {
            "items": {
              "$ref": "#/components/schemas/RespostaObjeto24Dto"
            },
            "type": "array"
          },
          "modo": {
            "enum": [
              "assistido",
              "autopiloto"
            ],
            "type": "string"
          },
          "oportunidadeId": {
            "nullable": true,
            "type": "string"
          },
          "pendencia": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RespostaObjeto33Dto"
              }
            ],
            "nullable": true,
            "type": "object"
          }
        },
        "required": [
          "id",
          "modo",
          "oportunidadeId",
          "mensagens",
          "pendencia",
          "criadoEm",
          "atualizadoEm"
        ],
        "type": "object"
      },
      "RespostaObjeto24Dto": {
        "properties": {
          "blocos": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/RespostaObjeto25Dto"
                },
                {
                  "$ref": "#/components/schemas/RespostaObjeto26Dto"
                },
                {
                  "$ref": "#/components/schemas/RespostaObjeto27Dto"
                },
                {
                  "$ref": "#/components/schemas/RespostaObjeto28Dto"
                },
                {
                  "$ref": "#/components/schemas/RespostaObjeto29Dto"
                }
              ]
            },
            "type": "array"
          },
          "conteudo": {
            "type": "string"
          },
          "dados": {
            "$ref": "#/components/schemas/RespostaObjeto30Dto"
          },
          "papel": {
            "enum": [
              "user",
              "assistant",
              "tool",
              "evento"
            ],
            "type": "string"
          },
          "tool": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "papel",
          "conteudo"
        ],
        "type": "object"
      },
      "RespostaObjeto25Dto": {
        "properties": {
          "text": {
            "type": "string"
          },
          "type": {
            "enum": [
              "text"
            ],
            "type": "string"
          }
        },
        "required": [
          "type",
          "text"
        ],
        "type": "object"
      },
      "RespostaObjeto26Dto": {
        "properties": {
          "id": {
            "type": "string"
          },
          "input": {
            "additionalProperties": {
              "oneOf": [
                {}
              ]
            },
            "type": "object"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "enum": [
              "tool_use"
            ],
            "type": "string"
          }
        },
        "required": [
          "type",
          "id",
          "name",
          "input"
        ],
        "type": "object"
      },
      "RespostaObjeto27Dto": {
        "properties": {
          "content": {
            "type": "string"
          },
          "is_error": {
            "enum": [
              false,
              true
            ],
            "type": "boolean"
          },
          "tool_use_id": {
            "type": "string"
          },
          "type": {
            "enum": [
              "tool_result"
            ],
            "type": "string"
          }
        },
        "required": [
          "type",
          "tool_use_id",
          "content"
        ],
        "type": "object"
      },
      "RespostaObjeto28Dto": {
        "properties": {
          "signature": {
            "type": "string"
          },
          "thinking": {
            "type": "string"
          },
          "type": {
            "enum": [
              "thinking"
            ],
            "type": "string"
          }
        },
        "required": [
          "type",
          "thinking",
          "signature"
        ],
        "type": "object"
      },
      "RespostaObjeto29Dto": {
        "properties": {
          "data": {
            "type": "string"
          },
          "type": {
            "enum": [
              "redacted_thinking"
            ],
            "type": "string"
          }
        },
        "required": [
          "type",
          "data"
        ],
        "type": "object"
      },
...
      "RespostaObjeto30Dto": {
        "properties": {
          "args": {
            "additionalProperties": {
              "oneOf": [
                {}
              ]
            },
            "type": "object"
          },
          "callId": {
            "type": "string"
          },
          "efeito": {
            "enum": [
              "leitura",
              "escrita",
              "entrega_externa"
            ],
            "type": "string"
          },
          "entrega": {
            "$ref": "#/components/schemas/RespostaObjeto31Dto"
          },
          "erro": {
            "type": "string"
          },
          "escopo": {
            "type": "string"
          },
          "etapa": {
            "enum": [
              1,
              3
            ],
            "type": "number"
          },
          "evento": {
            "enum": [
              "erro",
              "cancelado"
            ],
            "type": "string"
          },
          "jobId": {
            "type": "string"
          },
          "narracao": {
            "$ref": "#/components/schemas/RespostaObjeto32Dto"
          },
          "ok": {
            "enum": [
              false,
              true
            ],
            "type": "boolean"
          },
          "origem": {
            "enum": [
              "geracao_assincrona"
            ],
            "type": "string"
          },
          "resultado": {
            "oneOf": [
              {}
            ]
          }
        },
        "type": "object"
      },
      "RespostaObjeto31Dto": {
        "properties": {
          "destino": {
            "type": "string"
          },
          "texto": {
            "type": "string"
          },
          "tipo": {
            "type": "string"
          },
          "titulo": {
            "type": "string"
          }
        },
        "required": [
          "tipo",
          "titulo",
          "texto"
        ],
        "type": "object"
      },
      "RespostaObjeto32Dto": {
        "properties": {
          "degradacao": {
            "nullable": true,
            "type": "string"
          },
          "keywordsAusentes": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "keywordsAusentesIniciais": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "keywordsCobertas": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "keywordsEncontradas": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "pontosDeAtencao": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "scoreFinal": {
            "type": "number"
          },
          "scoreInicial": {
            "type": "number"
          },
          "veredicto": {
            "type": "string"
          }
        },
        "required": [
          "scoreInicial",
          "scoreFinal",
          "keywordsEncontradas",
          "keywordsAusentesIniciais",
          "pontosDeAtencao",
          "veredicto",
          "keywordsCobertas",
          "keywordsAusentes",
          "degradacao"
        ],
        "type": "object"
      },
      "RespostaObjeto33Dto": {
        "properties": {
          "args": {
            "additionalProperties": {
              "oneOf": [
                {}
              ]
            },
            "type": "object"
          },
          "callId": {
            "type": "string"
          },
          "efeito": {
            "enum": [
              "escrita"
            ],
            "type": "string"
          },
          "executando": {
            "enum": [
              false,
              true
            ],
            "type": "boolean"
          },
          "resumo": {
            "type": "string"
          },
          "tool": {
            "type": "string"
          }
        },
        "required": [
          "callId",
          "tool",
          "efeito",
          "args"
        ],
        "type": "object"
      },
      "RespostaObjeto34Dto": {
        "properties": {
          "destino": {
            "type": "string"
          },
          "texto": {
            "type": "string"
          },
          "tipo": {
            "enum": [
              "mensagem_recrutador"
            ],
            "type": "string"
          },
          "titulo": {
            "type": "string"
          }
        },
        "required": [
          "tipo",
          "titulo",
          "texto",
          "destino"
        ],
        "type": "object"
      },
      "RespostaObjeto35Dto": {
        "properties": {
          "respostas": {
            "items": {
              "$ref": "#/components/schemas/RespostaObjeto36Dto"
            },
            "type": "array"
          },
          "texto": {
            "type": "string"
          },
          "tipo": {
            "enum": [
              "resposta_formulario"
            ],
            "type": "string"
          },
          "titulo": {
            "type": "string"
          }
        },
        "required": [
          "tipo",
          "titulo",
          "respostas",
          "texto"
        ],
        "type": "object"
      },
      "RespostaObjeto36Dto": {
        "properties": {
          "campo": {
            "type": "string"
          },
          "texto": {
            "type": "string"
          }
        },
        "required": [
          "campo",
          "texto"
        ],
        "type": "object"
      },
...
      "RespostasFormularioDto": {
        "properties": {
          "campos": {
            "items": {
              "maxLength": 1000,
              "type": "string"
            },
            "type": "array"
          },
          "oportunidadeId": {
            "maxLength": 100,
            "type": "string"
          }
        },
        "required": [
          "oportunidadeId",
          "campos"
        ],
        "type": "object"
      },
...
      "UploadContextoDto": {
        "properties": {
          "arquivos": {
            "items": {
              "format": "binary",
              "type": "string"
            },
            "type": "array"
          },
          "historico": {
            "nullable": true,
            "oneOf": [
              {
                "type": "boolean"
              },
              {
                "enum": [
                  "true",
                  "false",
                  ""
                ],
                "type": "string"
              }
            ]
          }
        },
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "",
    "title": "prdal-careers-api",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/contexto/reindexar": {
      "post": {
        "operationId": "ContextoController_reindexar",
        "parameters": [],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaObjeto18Dto"
                }
              }
            },
            "description": ""
          },
          "default": {
            "description": ""
          }
        },
        "tags": [
          "Contexto"
        ]
      }
    },
    "/v1/contexto/status": {
      "get": {
        "operationId": "ContextoController_status",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaObjeto20Dto"
                }
              }
            },
            "description": ""
          },
          "default": {
            "description": ""
          }
        },
        "tags": [
          "Contexto"
        ]
      }
    },
    "/v1/contexto/upload": {
      "post": {
        "operationId": "ContextoController_upload",
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadContextoDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaObjeto19Dto"
                }
              }
            },
            "description": ""
          },
          "default": {
            "description": ""
          }
        },
        "tags": [
          "Contexto"
        ]
      }
    },
    "/v1/copiloto/chat": {
      "post": {
        "operationId": "CopilotoController_chatSse",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Copiloto"
        ]
      }
    },
    "/v1/copiloto/conversas": {
      "get": {
        "operationId": "CopilotoController_listarConversas",
        "parameters": [
          {
            "in": "query",
            "name": "oportunidadeId",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RespostaObjeto22Dto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          },
          "default": {
            "description": ""
          }
        },
        "tags": [
          "Copiloto"
        ]
      }
    },
    "/v1/copiloto/conversas/{id}": {
      "get": {
        "operationId": "CopilotoController_buscarConversa",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaObjeto23Dto"
                }
              }
            },
            "description": ""
          },
          "default": {
            "description": ""
          }
        },
        "tags": [
          "Copiloto"
        ]
      }
    },
    "/v1/copiloto/mensagem-recrutador": {
      "post": {
        "operationId": "CopilotoController_mensagemRecrutador",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MensagemRecrutadorDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaObjeto34Dto"
                }
              }
            },
            "description": ""
          },
          "default": {
            "description": ""
          }
        },
        "tags": [
          "Copiloto"
        ]
      }
    },
    "/v1/copiloto/respostas-formulario": {
      "post": {
        "operationId": "CopilotoController_respostasFormulario",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RespostasFormularioDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaObjeto35Dto"
                }
              }
            },
            "description": ""
          },
          "default": {
            "description": ""
          }
        },
        "tags": [
          "Copiloto"
        ]
      }
    },
...
  },
  "servers": [],
...
}